+++ src3_todos2/main.js
@@ -1,0 +1,15 @@
+/*
+入口JS
+ */
+import Vue from 'vue'
+import App from './App.vue'
+
+import './base.css'
+
+new Vue({
+  el: '#app',
+  components: {
+    App
+  },
+  template: '<App/>'
+})
